EY-4513: Flytter modell for simulering til utbetaling-model
